todoai-Flask/pyproject.toml
[tool.poetry]
name = "todoai-flask"
version = "0.1.0"
description = ""
authors = ["Victor Polyakov <polyakov.victor21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
jwt = "^1.3.1"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
